+++ src/slideshow/slideshow.js
@@ -16,16 +16,9 @@ import "./slideshow.scss";
  * @returns {JSX.Element} The component.
  */
 function Slideshow({ slide, content, run, slideDone, executionId }) {
-  const {
-    images,
-    imageDuration = 5000,
-    transitions,
-    animations,
-    // @TODO: Add when logo is available from theme
-    // , logoEnabled, logoSize, logoPosition
-  } = content;
+  const { images, imageDuration = 5000, transition, animation } = content;
   const [index, setIndex] = useState(0);
-  const fadeEnabled = transitions === "fade";
+  const fadeEnabled = transition === "fade";
   const fadeDuration = 1000;
   const [fade, setFade] = useState(false);
 
@@ -40,10 +33,6 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
 
   const timeoutRef = useRef(null);
   const fadeRef = useRef(null);
-
-  // @TODO: Get logo from theme.
-  // const logoImageUrl = null;
-  // const logoClasses = `logo ${logoPosition} ${logoSize}`;
 
   /**
    * A random function to simplify the code where random is used
@@ -102,7 +91,6 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
     ];
 
     const randomPercent = `${random(100) + 1}% ${random(100) + 1}%`;
-
     switch (animationType) {
       case "zoom-in-middle":
         return createAnimation(true);
@@ -122,11 +110,11 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
 
   // Setup animation
   useEffect(() => {
-    if (animations !== null) {
+    if (animation !== null) {
       // Adds the animation to the stylesheet. because there is an element of random, we cannot have it in the .scss file.
       const styleSheet = document.styleSheets[0];
       styleSheet.insertRule(
-        getCurrentAnimation(animations),
+        getCurrentAnimation(animation),
         styleSheet.cssRules.length
       );
     }
@@ -144,18 +132,18 @@ function Slideshow({ slide, content, run, slideDone, executionId }) {
 
     return imageStyle;
   };
-
   // Setup image progress.
   useEffect(() => {
     if (run) {
       if (imageUrls.length > 0) {
         timeoutRef.current = setTimeout(() => {
-          const newIndex = index + 1;
-
-          if (newIndex > imageUrls.length - 1) {
+          let newIndex = index + 1;
+          if (newIndex === imageUrls.length) {
+            newIndex = 0;
             // No more images to show.
             slideDone(slide);
-          } else if (fadeEnabled) {
+          }
+          if (fadeEnabled) {
             // Fade to next image.
             setFade(true);
             setAnimationIndex(newIndex);
@@ -252,11 +240,8 @@ Slideshow.propTypes = {
   content: PropTypes.shape({
     images: PropTypes.arrayOf(PropTypes.string),
     imageDuration: PropTypes.number,
-    logoEnabled: PropTypes.bool,
-    logoSize: PropTypes.string,
-    logoPosition: PropTypes.string,
-    animations: PropTypes.string,
-    transitions: PropTypes.string,
+    animation: PropTypes.string,
+    transition: PropTypes.string,
   }).isRequired,
   executionId: PropTypes.string.isRequired,
 };