+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Added logo to image-text.
 Fixed separator default styling.
 - [#109]https://github.com/os2display/display-templates/pull/109
 Make "hideGrid" work on calendar
+- [#108]https://github.com/os2display/display-templates/pull/110
+It slideshow restarts now and calls the done function. 
+rename transitions to transition, rename animations to animation.
+Changed "sekunder" to "millisekunder" on a label
 - [#112](https://github.com/os2display/display-templates/pull/112)
 Rename titles in screen layout six areas.